+++ ogma-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for ogma-core
 
+## [1.4.0] - 2024-05-21
+
+* Version bump 1.4.0 (#145).
+* Make ros command generate dockerfile (#136).
+* Map float and double to the same types in C++ (#138).
+
 ## [1.3.0] - 2024-03-21
 
 * Version bump 1.3.0 (#133).